requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-redis==5.4.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-django==4.7.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
